+++ localServer.js
@@ -1,14 +1,16 @@
 require('ts-node').register();
-const express = require('express');
-const bodyParser = require('body-parser');
-const pino = require('pino-http')({
-  level: 'info'
-}); // https://github.com/mcollina/pino#pinoopts-stream
+const fastify = require('fastify')
 
-
-const lambdaAdapter = (lambda, router) => {
+const lambdaAdapter = (lambda, app) => {
   // create a route that can talk to the lambda function
-  router.post('/', (req, res) => {
+  app.post('/', async (req, res) => {
+    try {
+    const result = await lambda(req.body);
+      res.send(result);
+    } catch(ex) {
+      res.code(406).send(ex);
+    }
+
     lambda(req.body, {}, (err, result) => {
       if (err) {
         return res.status(406).send(result);
@@ -19,39 +21,16 @@ const lambdaAdapter = (lambda, router) => {
   });
 };
 
-/**
- * Start the express server and accept port as a parameter and register a callback
- * @param {string} port
- * @param {func} callback
- */
-const createServer = (pathToLambda, callback) => {
-  const lambda = require(pathToLambda).default;
-
-  const app = express();
-  app.use(bodyParser.json());
-  app.use(pino); // instead of winston logger
-
-  app.use('*', (req, res, next) => {
-    req.logger = req.log; // so that it is available elsewhere.
-    next();
-  });
-
-  const router = express.Router({ // eslint-disable-line new-cap
-    caseSensitive: true
-  });
-  app.use(router);
-
-  lambdaAdapter(lambda, router);
-  return callback(app);
-};
-
-const runServer = (pathToLambda, options) => {
+const runServer = async (pathToLambda, options) => {
   const localOptions = options || {};
 
-  createServer(pathToLambda, (app) => {
-    const listener = app.listen(localOptions.localPort || 9000, () => {
-      console.log(`Server started on port ${listener.address().port}`);
-    });
+  const lambda = require(pathToLambda).default;
+
+  const app = fastify();
+  lambdaAdapter(lambda, app);
+
+  await app.listen(localOptions.localPort || 9000, () => {
+    console.log(`Server started on port ${app.server.address().port}`);
   });
 };